Added endpoint to translate mentions before submitting to chain, edit resolver
