+++ .chain/graphql-schema.json
@@ -2045,6 +2045,20 @@
                     "ofType": null
                   }
                 }
+              },
+              {
+                "defaultValue": null,
+                "description": "",
+                "name": "raw",
+                "type": {
+                  "kind": "NON_NULL",
+                  "name": "",
+                  "ofType": {
+                    "kind": "SCALAR",
+                    "name": "bool",
+                    "ofType": null
+                  }
+                }
               }
             ],
             "deprecationReason": "",
@@ -2538,18 +2552,6 @@
             }
           },
           {
-            "args": [],
-            "deprecationReason": "",
-            "description": "",
-            "isDeprecated": false,
-            "name": "unreadNotificationsCount",
-            "type": {
-              "kind": "OBJECT",
-              "name": "UnreadNotificationEventsCount",
-              "ofType": null
-            }
-          },
-          {
             "args": [
               {
                 "defaultValue": null,
@@ -2578,6 +2580,18 @@
                 "name": "Story",
                 "ofType": null
               }
+            }
+          },
+          {
+            "args": [],
+            "deprecationReason": "",
+            "description": "",
+            "isDeprecated": false,
+            "name": "unreadNotificationsCount",
+            "type": {
+              "kind": "OBJECT",
+              "name": "UnreadNotificationEventsCount",
+              "ofType": null
             }
           },
           {
