+++ x/db/comment.go
@@ -1,12 +1,6 @@
 package db
 
 import (
-	"fmt"
-	"net/url"
-	"path"
-	"strings"
-
-	"github.com/gernest/mention"
 	"github.com/kelseyhightower/envconfig"
 )
 
@@ -53,54 +47,6 @@ func (c *Client) CommentsByArgumentID(argumentID int64) ([]Comment, error) {
 	return transformedComments, nil
 }
 
-// replace @cosmosaddr with profile link [@username](https://app.trustory.io/profile/cosmosaddr)
-func (c *Client) replaceAddressesWithProfileURLs(config ChainConfig, body string) (string, error) {
-	profileURLPrefix := path.Join(config.Host, "profile")
-	profileURLsByAddress, err := c.mapAddressesToProfileURLs(config, body, profileURLPrefix)
-	if err != nil {
-		return "", err
-	}
-	for address, profileURL := range profileURLsByAddress {
-		body = strings.ReplaceAll(body, fmt.Sprintf("@%s", address), profileURL)
-	}
-
-	return body, nil
-}
-
-func (c *Client) mapAddressesToProfileURLs(config ChainConfig, body string, profileURLPrefix string) (map[string]string, error) {
-	profileURLsByAddress := map[string]string{}
-	addresses := parseMentions(body)
-	for _, address := range addresses {
-		twitterProfile, err := c.TwitterProfileByAddress(address)
-		if err != nil {
-			return profileURLsByAddress, err
-		}
-		if twitterProfile == nil {
-			profileURLsByAddress[address] = address
-			continue
-		}
-		profileURLString := path.Join(profileURLPrefix, twitterProfile.Address)
-		profileURL, err := url.Parse(profileURLString)
-		if err != nil {
-			return profileURLsByAddress, err
-		}
-
-		httpPrefix := "http://"
-		if config.LetsEncryptEnabled == true {
-			httpPrefix = "https://"
-		}
-		markdownProfileURL := fmt.Sprintf("[@%s](%s%s)", twitterProfile.Username, httpPrefix, profileURL)
-		profileURLsByAddress[address] = markdownProfileURL
-	}
-
-	return profileURLsByAddress, nil
-}
-
-// extract @mentions from text and return as slice
-func parseMentions(body string) []string {
-	return mention.GetTagsAsUniqueStrings('@', body, ' ', '\n', '\r')
-}
-
 // AddComment adds a new comment to the comments table
 func (c *Client) AddComment(comment *Comment) error {
 	transformedBody, err := c.replaceUsernamesWithAddress(comment.Body)
@@ -114,28 +60,6 @@ func (c *Client) AddComment(comment *Comment) error {
 	}
 
 	return nil
-}
-
-// replace @usernames with @cosmosaddr
-func (c *Client) replaceUsernamesWithAddress(body string) (string, error) {
-	addressByUsername := map[string]string{}
-	usernames := parseMentions(body)
-	for _, username := range usernames {
-		twitterProfile, err := c.TwitterProfileByUsername(username)
-		if err != nil {
-			return body, err
-		}
-		if twitterProfile == nil {
-			addressByUsername[username] = username
-		} else {
-			addressByUsername[username] = twitterProfile.Address
-		}
-	}
-	for username, address := range addressByUsername {
-		body = strings.ReplaceAll(body, username, address)
-	}
-
-	return body, nil
 }
 
 // CommentsParticipantsByArgumentID gets the list of users participating on a argument thread.
