+++ x/truapi/resolver.go
@@ -99,7 +99,7 @@ func (ta *TruAPI) storiesResolver(_ context.Context, q app.QueryByCategoryIDPara
 	return filteredStories
 }
 
-func (ta *TruAPI) argumentResolver(_ context.Context, q app.QueryByIDParams) argument.Argument {
+func (ta *TruAPI) argumentResolver(_ context.Context, q app.QueryArgumentByID) argument.Argument {
 	res := ta.RunQuery(
 		path.Join(argument.QueryPath, argument.QueryArgumentByID),
 		app.QueryByIDParams{ID: q.ID},
@@ -115,6 +115,16 @@ func (ta *TruAPI) argumentResolver(_ context.Context, q app.QueryByIDParams) arg
 	if err != nil {
 		panic(err)
 	}
+	// check if raw argument was passed
+	if q.Raw {
+		return *argument
+	}
+
+	body, err := ta.DBClient.TranslateToUsersMentions(argument.Body)
+	if err != nil {
+		panic(err)
+	}
+	argument.Body = body
 
 	return *argument
 }
